+++ test/dotnet.Tests/CommandTests/ToolManifestFinderTests.cs
@@ -19,14 +19,14 @@ using System.Linq;
 
 namespace Microsoft.DotNet.Tests.Commands
 {
-    public class ToolManifestTests
+    public class ToolManifestFinderTests
     {
         private readonly IFileSystem _fileSystem;
         private readonly List<ToolManifestPackage> _defaultExpectedResult;
         private readonly string _testDirectoryRoot;
         private const string _manifestFilename = "dotnet-tools.json";
 
-        public ToolManifestTests()
+        public ToolManifestFinderTests()
         {
             _fileSystem = new FileSystemMockBuilder().UseCurrentSystemTemporaryDirectory().Build();
             _testDirectoryRoot = _fileSystem.Directory.CreateTemporaryDirectory().DirectoryPath;
@@ -365,120 +365,6 @@ namespace Microsoft.DotNet.Tests.Commands
 
             Action a = () => toolManifest.TryFind(new ToolCommandName("dotnetsay"), out var result);
             a.ShouldThrow<ToolManifestException>();
-        }
-
-        [Fact]
-        public void GivenManifestFileOnSameDirectoryItCanAddEntryToIt()
-        {
-            string manifestFile = Path.Combine(_testDirectoryRoot, _manifestFilename);
-            _fileSystem.File.WriteAllText(manifestFile, _jsonContent);
-
-            var toolManifestFileEditor = new ToolManifestEditor(_fileSystem);
-
-            toolManifestFileEditor.Add(new FilePath(manifestFile),
-                new PackageId("new-tool"),
-                NuGetVersion.Parse("3.0.0"),
-                new[] { new ToolCommandName("newtool") });
-
-            _fileSystem.File.ReadAllText(manifestFile).Should().Be(
-@"{
-  ""version"": 1,
-  ""isRoot"": true,
-  ""tools"": {
-    ""t-rex"": {
-      ""version"": ""1.0.53"",
-      ""commands"": [
-        ""t-rex""
-      ]
-    },
-    ""dotnetsay"": {
-      ""version"": ""2.1.4"",
-      ""commands"": [
-        ""dotnetsay""
-      ]
-    },
-    ""new-tool"": {
-      ""version"": ""3.0.0"",
-      ""commands"": [
-        ""newtool""
-      ]
-    }
-  }
-}");
-        }
-
-        [Fact]
-        public void GivenManifestFileOnSameDirectoryWhenAddingTheSamePackageIdToolItThrows()
-        {
-            string manifestFile = Path.Combine(_testDirectoryRoot, _manifestFilename);
-            _fileSystem.File.WriteAllText(manifestFile, _jsonContent);
-
-            var toolManifestFileEditor = new ToolManifestEditor(_fileSystem);
-
-            PackageId packageId = new PackageId("dotnetsay");
-            NuGetVersion nuGetVersion = NuGetVersion.Parse("3.0.0");
-            Action a = () => toolManifestFileEditor.Add(new FilePath(manifestFile),
-                packageId,
-                nuGetVersion,
-                new[] {new ToolCommandName("dotnetsay")});
-
-
-            var expectedString = string.Format(
-                LocalizableStrings.ManifestPackageIdCollision,
-                packageId.ToString(),
-                nuGetVersion.ToNormalizedString(),
-                manifestFile,
-                packageId.ToString(),
-                "2.1.4");
-
-            a.ShouldThrow<ToolManifestException>()
-                .And.Message.Should().Contain(expectedString);
-
-            _fileSystem.File.ReadAllText(manifestFile).Should().Be(_jsonContent);
-        }
-
-        [Fact]
-        public void GivenManifestFileOnSameDirectoryWhenAddingTheSamePackageIdSameVersionSameCommandsItDoesNothing()
-        {
-            string manifestFile = Path.Combine(_testDirectoryRoot, _manifestFilename);
-            _fileSystem.File.WriteAllText(manifestFile, _jsonContent);
-
-            var toolManifestFileEditor = new ToolManifestEditor(_fileSystem);
-
-            PackageId packageId = new PackageId("dotnetsay");
-            NuGetVersion nuGetVersion = NuGetVersion.Parse("2.1.4");
-            Action a = () => toolManifestFileEditor.Add(new FilePath(manifestFile),
-                packageId,
-                nuGetVersion,
-                new[] { new ToolCommandName("dotnetsay") });
-
-            a.ShouldNotThrow();
-
-            _fileSystem.File.ReadAllText(manifestFile).Should().Be(_jsonContent);
-        }
-
-        [Fact]
-        public void GivenAnInvalidManifestFileOnSameDirectoryWhenAddItThrows()
-        {
-            string manifestFile = Path.Combine(_testDirectoryRoot, _manifestFilename);
-            _fileSystem.File.WriteAllText(manifestFile, _jsonWithInvalidField);
-
-            var toolManifestFileEditor = new ToolManifestEditor(_fileSystem);
-
-            PackageId packageId = new PackageId("dotnetsay");
-            NuGetVersion nuGetVersion = NuGetVersion.Parse("3.0.0");
-            Action a = () => toolManifestFileEditor.Add(new FilePath(manifestFile),
-                packageId,
-                nuGetVersion,
-                new[] {new ToolCommandName("dotnetsay")});
-
-            a.ShouldThrow<ToolManifestException>()
-                .And.Message.Should().Contain(
-                    string.Format(LocalizableStrings.InvalidManifestFilePrefix,
-                        manifestFile,
-                        string.Empty));
-
-            _fileSystem.File.ReadAllText(manifestFile).Should().Be(_jsonWithInvalidField);
         }
 
         [Fact]
